Moved loadTomlConfigFile() to fileHelper.ts, fixed the imports in index.ts
